cambios

+++ src/main/java/com/cetys/loading/model/Audit.java
@@ -1,19 +1,25 @@
 package com.cetys.loading.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@Table(name="audit")
+@Table(name = "audit")
 @Getter
 @Setter
 public class Audit extends BaseEntity {
 
     @Id
-    @Column(name="audit_id")
+    @Column(name = "audit_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int auditId;
+    Long auditId;
 
     @ManyToOne
     Subarea subarea;
