+++ src/main/java/com/cetys/loading/model/Org.java
@@ -1,20 +1,25 @@
 package com.cetys.loading.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@Table(name="org")
+@Table(name = "org")
 @Getter
 @Setter
 public class Org extends BaseEntity {
 
     @Id
-    @Column(name="org_id")
+    @Column(name = "org_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int org_id;
+    Long org_id;
 
-    @Column(name="name")
+    @Column(name = "name")
     String name;
 }
