+++ src/main/java/com/cetys/loading/model/Subarea.java
@@ -1,23 +1,31 @@
 package com.cetys.loading.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@Table(name="subarea")
+@Table(name = "subarea")
 @Getter
 @Setter
 public class Subarea extends BaseEntity {
     @Id
-    @Column(name="subarea_id")
+    @Column(name = "subarea_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int subarea_id;
+    Long subarea_id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "area_id")
     Area area;
 
-    @Column(name="name")
+    @Column(name = "name")
     String name;
 }
