+++ src/main/java/com/cetys/loading/model/User.java
@@ -1,25 +1,28 @@
 package com.cetys.loading.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @Entity
-@Table(name="app_user")
+@Table(name = "app_user")
 @Getter
 @Setter
 public class User extends BaseEntity {
 
     @Id
-    @Column(name="user_id")
+    @Column(name = "user_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int userId;
+    Long userId;
 
-    @Column(name="name")
+    @Column(name = "name")
     String name;
 
-    @Column(name="email")
+    @Column(name = "email")
     String email;
 }
