+++ src/main/java/com/cetys/loading/model/UserRole.java
@@ -1,32 +1,37 @@
 package com.cetys.loading.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.annotations.Fetch;
-
-import java.time.LocalDateTime;
 
 @Entity
-@Table(name="app_user_role")
+@Table(name = "app_user_role")
 @Getter
 @Setter
 public class UserRole extends BaseEntity {
 
     @Id
-    @Column(name="user_role_id")
+    @Column(name = "user_role_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int userRoleId;
+    Long userRoleId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-            @JoinColumn(name="user_id")
+    @JoinColumn(name = "user_id")
     User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="org_id")
+    @JoinColumn(name = "org_id")
     Org org;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="allowed_subarea_id")
+    @JoinColumn(name = "allowed_subarea_id")
     Subarea allowedSubarea;
 }
